requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.4.0
beautifulsoup4>=4.10.0
html2markdown>=0.1.7
python-frontmatter>=1.0.0
python-slugify>=8.0.0
tqdm>=4.62.0
sentence-transformers>=2.2.0
requests>=2.27.0
python-dotenv>=1.0.0

# PDF extraction backends
PyPDF2>=2.10.0
pdfplumber>=0.7.0

# Optional dependencies
chardet>=5.0.0
scikit-learn>=1.0.0  # For additional vector operations

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0